Struct deltae::DeltaE

pub struct DeltaE { /* fields omitted */ }

The measured difference between two colors

There are many different methods of calculating color difference. Different methods have a specific purpose, mainly in determining the level of tolerance for describing the difference between two colors. Regardless of the DEMethod used, DeltaE is always calculated based on the LabValues of the two colors.

Implementations

impl DeltaE

pub fn new<A, B>(a: A, b: B, method: DEMethod) -> DeltaE where
A: Delta,
B: Delta,

New DeltaE from two colors and a DEMethod.

use deltae::{LabValue, DeltaE, DEMethod::DE2000};

let lab0 = LabValue::new(89.73, 1.88, -6.96).unwrap();
let lab1 = LabValue::new(95.08, -0.17, -10.81).unwrap();
let de0 = DeltaE::new(&lab0, &lab1, DE2000);
assert_eq!(de0, 5.316941);

pub fn with_method(self, method: DEMethod) -> Self

Recalculate DeltaE with another DEMethod

use deltae::{Delta, DeltaE, LabValue, DEMethod};

let lab0 = LabValue::new(89.73, 1.88, -6.96).unwrap();
let lab1 = LabValue::new(95.08, -0.17, -10.81).unwrap();
let de2000 = lab0.delta(lab1, DEMethod::DE2000);
let de1976 = de2000.with_method(DEMethod::DE1976);
assert_eq!(de1976, 6.902716);

pub fn method(&self) -> &DEMethod

Return a reference the DeltaE method used in the calculation

pub fn value(&self) -> &f32

Return a reference to the DeltaE value

pub fn reference(&self) -> &LabValue

Return a reference to the reference LabValue used in the calculation. A reference color is the base color to which the sample color is being compared.

pub fn sample(&self) -> &LabValue

Return a reference to the sample LabValue used in the calculation. A sample color is the color being compared to the reference color.