Trait deltae::Round

pub trait Round {
    fn round_to(self, places: i32) -> Self;
}

Trait for rounding values to a number of decimal places

Required methods

fn round_to(self, places: i32) -> Self

Rounds the value to a number of decimal places

Loading content...